requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
pydub==0.25.1
regex==2024.11.6
requests==2.32.3
setuptools==77.0.1
SpeechRecognition==3.14.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
